+++ packages/moonstone/CHANGELOG.md
@@ -8,6 +8,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/IncrementSlider` prop `children` which was no longer supported for setting the tooltip (since 2.0.0-beta.1)
 
+### Fixed
+
+- `moonstone/Scroller` ordering of logic for Scroller focus to check focus possibilities first then go to fallback at the top of the container
+- `moonstone/Scroller` to scroll by page when focus was at the edge of the viewport
+
 ## [2.0.0-beta.5] - 2018-05-29
 
 ### Removed