+++ packages/moonstone/CHANGELOG.md
@@ -4,6 +4,9 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [Unreleased]
 
+### Added
+- `moonstone/Scrollable` props `spotlightPaging` to determine how spotlight moves when page up/down are pressed
+
 ### Fixed
 - `moonstone/Scroller` ordering of logic for Scroller focus to check focus possibilities first then go to fallback at the top of the container
 - `moonstone/Scroller` to scroll by page when focus was at the edge of the viewport