+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 - `moonstone/Scroller` ordering of logic for Scroller focus to check focus possibilities first then go to fallback at the top of the container
 - `moonstone/Scroller` to scroll by page when focus was at the edge of the viewport
+- `moonstone/VideoPlayer` to show thumbnail only when focused on slider
 - `moonstone/VideoPlayer` to not hide title and info section when showing more components
 - `moonstone/VideoPlayer` to select a position in slider to seek in 5-way mode