+++ packages/moonstone/CHANGELOG.md
@@ -15,6 +15,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/VideoPlayer` to show thumbnail only when focused on slider
 - `moonstone/VideoPlayer` to not hide title and info section when showing more components
 - `moonstone/VideoPlayer` to select a position in slider to seek in 5-way mode
+- `moonstone/ToggleButton` to have correct padding and orientation for RTL
 
 ## [2.0.0-beta.5] - 2018-05-29