+++ packages/moonstone/CHANGELOG.md
@@ -7,9 +7,18 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Added
 
 - `moonstone/VideoPlayer.MediaControls` property `playerNode`
+### Removed
+
+- `moonstone/IncrementSlider` prop `children` which was no longer supported for setting the tooltip (since 2.0.0-beta.1)
 
 ### Fixed
+
 - `moonstone/Scroller` ordering of logic for Scroller focus to check focus possibilities first then go to fallback at the top of the container
+- `moonstone/Scroller` to scroll by page when focus was at the edge of the viewport
+- `moonstone/VideoPlayer` to show thumbnail only when focused on slider
+- `moonstone/VideoPlayer` to not hide title and info section when showing more components
+- `moonstone/VideoPlayer` to select a position in slider to seek in 5-way mode
+- `moonstone/ToggleButton` to have correct padding and orientation for RTL
 
 ## [2.0.0-beta.5] - 2018-05-29
 
@@ -42,6 +51,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Button` and `moonstone/IconButton` class name `small` to the list of allowed `css` overrides
 - `moonstone/VideoPlayer.MediaControls` property `onClose` to handle back key
 - `moonstone/ProgressBar` prop `highlighted` for when the UX needs to call special attention to a progress bar
+
+### Changed
+
+- `moonstone/VideoPlayer` to disable media slider when source is unavailable
 
 ### Fixed